Create problem.md
in a given array of n elements, and for the given value of sum of 2 numbers -> sum, find out whether there exist two elements in array such that their sum equals the given value of sum.  
this is the brute force solution with time complexity as O(n^2) and space complexity as O(1).  
# example:  
arr=[1,2,3,4,5] sum=5 -> true.