Create problem.md
in a given array of n elements, and for the given value of sum of 2 numbers -> sum, find out whether there exist two elements in array such that their sum equals the given value of sum and return the indices of those elements. 

this is the better solution with time complexity as O(n) and space complexity as O(n).  
# example:  
arr=[1,2,3,4,5] sum=5 -> (1,2) (0,3).