Create problem.md
in a given array of n elements where n is always even, there are exactly half of the elements that are positive and the other half are negative.  
rearrange the array elements such that the order of the elements of a particular sign remain the same as input array but the elements are now having alternate signs.  
# example :
[1,3,-4,-5,6,-7] -> [1,-4,3,-5,6,-7]