Create problem.md
an element in an array is said to be leader if there exists no element on the right side of it which is greater.  
from the given array of n elements, you need to find all the leaders of the array and return them.  
this is the optimal solution approach and the time complexity is O(n) and space complexity is O(1).
# example:
[10 22 12 3 0 6] -> [22 12 6]