Create problem.md
Given a sorted array of nums and an integer x, write a program to find the upper bound of x.

The upper bound of x is defined as the smallest index i such that nums[i] > x.

If no such index is found, return the size of the array.  
 
Examples:  
Input : n= 4, nums = [1,2,2,3], x = 2  
Output:3  
Explanation:  
Index 3 is the smallest index such that arr[3] > x.