Create problem.md
Given a sorted array nums and an integer x. Find the floor and ceil of x in nums. The floor of x is the largest element in the array which is smaller than or equal to x. The ceiling of x is the smallest element in the array greater than or equal to x. If no floor or ceil exists, output -1.

Examples:  
Input : nums =[3, 4, 4, 7, 8, 10], x= 5  
Output: 4 7  
Explanation: The floor of 5 in the array is 4, and the ceiling of 5 in the array is 7.